Escape the generated URL for output as an attribute
This is already handled mostly in `duplicate_post_get_clone_post_link()`,
but can be filtered afterward. Escape as late as possible. I might
use `esc_url()` here instead, but the display context already sets
`&amp`, which would then turn into `&#038` if we ran `esc_url()`
on it.

This isn't critical, but nice to have.