When merging back, don't override original post date
When merging back to original post, original post date should not be overridden by duplicate post date. This is a big issue in editorial workflow and make difficult to properly track post updates. The proposed change fixes this issue by checking if we are merging post and then assigning post date to original.